test-projects/py-ml/pyproject.toml
[project]
name = "py-ml"
version = "0.1.0"
description = "Python ML pipeline for grit coordination testing"
requires-python = ">=3.11"
dependencies = [
    "numpy>=1.26",
    "pandas>=2.1",
    "scikit-learn>=1.3",
    "matplotlib>=3.8",
    "pydantic>=2.5",
    "pyyaml>=6.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "pytest>=7.4",
    "pytest-cov>=4.1",
    "ruff>=0.1",
]
